+++ lib/simple_form/inputs/base.rb
@@ -211,9 +211,14 @@ module SimpleForm
       def set_input_classes(wrapper_options)
         wrapper_options = wrapper_options.dup
         error_class     = wrapper_options.delete(:error_class)
+        valid_class     = wrapper_options.delete(:valid_class)
 
         if error_class.present? && has_errors?
           wrapper_options[:class] = "#{wrapper_options[:class]} #{error_class}"
+        end
+
+        if valid_class.present? && valid?
+          wrapper_options[:class] = "#{wrapper_options[:class]} #{valid_class}"
         end
 
         wrapper_options